requirements.txt
pandas
google-api-python-client
google-auth
pytest
requests
logging
httpx
openai
python-dotenv